aut updatetd

+++ src/products/products.service.ts
@@ -1,6 +1,7 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { User } from 'src/user/entity/user.model';
 
 import { Product } from './product.model';
 @Injectable()
@@ -8,13 +9,18 @@ export class ProductsService {
 
   constructor (@InjectModel('Product') private readonly productModel : Model<Product>) {}
   
-  async insertProduct(product : Product) {
+  async insertProduct(product : Product , user : User) {
+  if(user.role == 'admin') {
+
     var obj = JSON.parse(JSON.stringify(product));
     const newProduct = new this.productModel(
       obj
     );
     newProduct.save();
     return newProduct;
+  }
+
+    throw new UnauthorizedException("ONLY ADMINS CAN ADD PRODUCTS")
    }
 
     async getProducts() {
@@ -59,7 +65,27 @@ export class ProductsService {
         return  product ;
     }
 
-     async   removeProduct(id: string ) : Promise<any> {
+    /////////////////////////////////////////
+
+     async   findProductByCategory(category: string ) : Promise<Product> {
+      let product ;
+     try {
+         product = await this.productModel.find({form: category}).exec()
+     } catch (e) {
+      throw new NotFoundException('Could Not Found this product ');
+     }
+  
+     if (!product) {
+      throw new NotFoundException('Could Not Found this product ');
+     }
+          return  product ;
+      }
+  
+
+     async   removeProduct(id: string , user :User) : Promise<any> {
+       console.log(user)
+     if (user.role == 'admin') 
+     {
       let product ;
      try {
          product = await this.productModel.findByIdAndDelete(id).exec();
@@ -71,10 +97,18 @@ export class ProductsService {
      if (!product) {
       throw new NotFoundException('Could Not Found this product ');
      }
+     return (product);
+    }
+     throw new UnauthorizedException("ONLY ADMINS CAN DELETE PRODUCTS")
       
       }
 
-      async   updateProduct( newProduct : Product) : Promise<any> {
+      async   updateProduct( newProduct : Product , user : User) : Promise<any> {
+   
+        if (user.role == 'admin') {
+      if(!newProduct.id) {
+        throw new NotFoundException("WE NEED ID RIGHT HERE PLEASE") ;
+      }
         let product ;
        try {
            product = await this.productModel.updateOne({_id:newProduct.id},newProduct).exec();
@@ -87,9 +121,12 @@ export class ProductsService {
        if (!product) {
         throw new NotFoundException('Could Not Found this product ');
        }
+       return (product) ;
         
         }
-
+        throw new UnauthorizedException("ONLY ADMINS CAN UPDATE PRODUCTS") ;
+      }
+       
 
   }
